Move RerunableSqlTableJournal to Extension class
